requirements.txt
numpy
opencv-python
transformers==4.31.0
taming-transformers
diffusers
xformers==0.0.27post1

accelerate
pytorch_lightning
einops
omegaconf
torchmetrics
opencv-python-headless
scipy
matplotlib
open-clip-torch==2.24.0
lpips
gradio==3.24.0
chardet
facexlib
pyiqa
pydantic==1.10.11
packaging
sentencepiece
fastapi==0.100.1
markdown2==2.4.10
openai==0.27.8
pycocotools==2.0.6
ray==2.6.1
shortuuid==1.0.11
tqdm==4.64.1
uvicorn==0.23.2
scipy==1.11.2
bitsandbytes==0.41.1
